+++ mona/core/file_server/ISO9660/ISO9660FileSystem.h
@@ -17,7 +17,7 @@ public:
 public:
     virtual int initialize();
     virtual int lookup(Vnode* diretory, const std::string& file, Vnode** found, int type);
-    virtual int open(Vnode* file, int mode);
+    virtual int open(Vnode* file, intptr_t mode);
     virtual int create(Vnode* dir, const std::string& file);
     virtual int read(Vnode* file, struct io::Context* context);
     virtual int write(Vnode* file, struct io::Context* context);
